requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Redis
redis==5.0.1

# Celery
celery[redis]==5.3.6
flower==2.0.1

# HTTP clients
httpx==0.26.0
websockets==12.0
msgpack==1.0.7

# Data processing
pandas==2.1.4
numpy==1.26.3
scipy==1.11.4
xgboost==2.0.3
scikit-learn>=1.3.0

# Utilities
python-dotenv==1.0.0
structlog==24.1.0
tenacity==8.2.3
pyyaml==6.0.1

# Polymarket Trading
py-clob-client==0.18.0
PySocks==1.7.1  # SOCKS proxy support for requests

# LLM
openai==1.51.0

# Fuzzy string matching
rapidfuzz>=3.0.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
